pyproject.toml
[tool.poetry]
name = "task-user-shenanigans"
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.0.6"
django-dramatiq = "^0.11.6"
psycopg = {extras = ["binary"], version = "^3.1.19"}
pika = "^1.3.2"

[tool.poetry.group.dev.dependencies]
ruff = "^0.4.4"

[tool.ruff]
line-length = 120
src = ["src"]
exclude = ["migrations"]

[tool.ruff.lint]
select = [
    "F",
    "E",
    "W",
    "C90",
    "I",
    "N",
    "UP",
    "DJ",
    "EM",
    "Q",
    "RUF",
]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
